Port in octodns.processor.spf
